Feat+Fix: 테스트 모델링 및 애니메이션 추가, 총 발사 메커니즘 변경
1. 테스트로 사용될 모델링 추가
2. 테스트 모델링에 대한 애니메이터 및 애니메이터 제어 스크립트 추가(TestPlayer.cs인데, 테스트용으로만 쓰십쇼)
3. 오브젝트에 대한 emission 메커니즘 (팔레트방식)으로 변경해보는 중(진행중임)

+++ Assets/Scripts/Taeseung/Weapon/Bomb/SubWeapon_BombManager.cs
@@ -26,10 +26,8 @@ public class SubWeapon_BombManager: SubWeaponManager, WeaponInterface
 
     private void Update()
     {
-        if (Input.GetMouseButtonDown(0) && _weaponCount > 0)
-        {
+        if (Input.GetMouseButtonDown(0) && _weaponCount > 0){
             StartAttack();
-
         }
     }
 
@@ -79,12 +77,12 @@ public class SubWeapon_BombManager: SubWeaponManager, WeaponInterface
         }
     }
 
-    public void SetWeaponGauge(float newval)
+    public void SetWeaponGauge(int newval)
     {
         //not need
     }
 
-    public float GetWeaponGauge()
+    public int GetWeaponGauge()
     {
         return 0;//not need
     }
